RF: test_pets_verafes separated by method
